requirements.txt
asgiref==3.8.1
Django==5.0.3
django-jazzmin==2.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
PyJWT==2.6.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
shortuuid==1.0.13
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
